Refactor async operators to their own files
