+++ lib/async/accumulate.ts
@@ -1,41 +1,4 @@
-import type { Callback, UnaryOperator } from "..";
-import { iterate, type Accumulator } from ".";
-
-function each<T>(iterable: AsyncIterable<T>, callback: Callback<T>): UnaryOperator<void, Promise<void>> {
-    const iterator = iterate(iterable);
-
-    (async function() {
-        let result = await iterator.next();
-
-        while (!result.done) {
-            callback(result.value);
-            result = await iterator.next();
-        }
-    })();
-
-    return async () => {
-        await iterator.return?.();
-    };
-}
-
-function collect<T>(iterable: AsyncIterable<T>): UnaryOperator<void, Array<T>> {
-    const iterator = iterate(iterable);
-    const results = new Array<T>();
-
-    (async function() {
-        let result = await iterator.next();
-
-        while (!result.done) {
-            results.push(result.value);
-            result = await iterator.next();
-        }
-    })();
-
-    return (): Array<T> => {
-        iterator.return?.();
-        return results;
-    };
-}
+import { type Accumulator, iterate } from ".";
 
 async function* accumulate<T, R>(iterable: AsyncIterable<T>, accumulator: Accumulator<T, R>, value: R): AsyncIterator<R> {
     const iterator = iterate(iterable);
@@ -48,8 +11,8 @@ async function* accumulate<T, R>(iterable: AsyncIterable<T>, accumulator: Accumu
             result = await iterator.next();
         }
     } finally {
-        iterator.return?.();
+        await iterator.return?.();
     }
 }
 
-export { each, collect, accumulate };
+export default accumulate;
